requirements.txt
numpy
pandas
opencv-python
Pillow
matplotlib
scikit-learn
torch
torchvision
torchaudio
easyocr
pytesseract
editdistance
tqdm
albumentations
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
beautifulsoup4